finished copying all the datasource methods in deck, card, and relations into the quizdatasource. still need to safe delete and make changes in other files
